Renamed old get method
